requirements.txt
ephem
python-telegram-bot[socks]
flask
requests